pyproject.toml
[project]
name = "pay-regulations-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.59.0",
    "faiss-cpu>=1.11.0.post1",
    "fastapi>=0.116.1",
    "google-generativeai>=0.8.5",
    "google-cloud-bigquery>=3.25.0",
    "langchain>=0.3.26",
    "langchain-community>=0.3.27",
    "langchain-google-genai>=2.0.10",
    "matplotlib>=3.10.3",
    "numpy>=2.3.1",
    "openpyxl>=3.1.5",
    "paddleocr>=3.1.0",
    "paddlepaddle>=3.1.0",
    "pandas>=2.3.1",
    "pillow>=11.3.0",
    "pypdf2>=3.0.1",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.1.1",
    "seaborn>=0.13.2",
    "setuptools>=80.9.0",
    "streamlit>=1.47.0",
    "tesseract>=0.1.3",
    "uvicorn[standard]>=0.35.0",
    "vertexai>=1.71.1",
    "watchdog>=6.0.0",
]